fix to run on new release of SiEPIC EBeam PDK
Was previously not working, fixed to run and build a new image whenever a new release of SiEPIC EBeam PDK is released. Unfortunately there is no straight forward way to trigger the action when a release is made to another repo (i.e SiEPIC Tools), this will need to be further investigated.