EBeam:"contra_directional_coupler" model update
Being updated for the OFC SC432 tutorial.
Issue:  The Lumerical Interconnect component models had Simulation parameters in addition to the usual Device parameters model, which need to be entered. This is because our data filenames include the wavelength start, wavelength end, and number of points. This improves the model script to pick a file that only based on the physical Device parameters, if the exact one is not found. Issue a message (can be disabled).